requirements.txt
# ============================================================
# IMPORTANT: Python 3.12 is required for this project.
#
# Reasons:
#   - pandas-ta pins numba==0.61.2, which only builds on Python <=3.12
#   - torch >=2.6 fails on Windows 10 (WinError 1114, c10.dll init)
#
# After running `pip install -r requirements.txt`, also run:
#   pip install "torch==2.5.0" --index-url https://download.pytorch.org/whl/cpu
#
# Setup: py -3.12 -m venv .venv && .venv\Scripts\activate
# ============================================================

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0
click>=8.1

# Data
yfinance>=0.2.40
pandas>=2.1.0
polars>=0.20.0
pyarrow>=15.0.0
numpy>=1.26.0

# Indicators
# NOTE: pandas-ta pins numba==0.61.2. This works on Python 3.12 (the required version).
# Do not add a numba version override here — it will cause a pip resolution conflict.
pandas-ta>=0.3.14b

# Sentiment (FinBERT)
transformers>=4.40.0
# torch: pip's default index picks the latest (2.11+) which fails on Windows 10 with WinError 1114.
# Install CPU wheel manually AFTER pip install -r requirements.txt:
#   pip install "torch==2.5.0" --index-url https://download.pytorch.org/whl/cpu
# torch 2.5.0+cpu is confirmed working on Python 3.12 + Windows 10.
torch>=2.2.0,<2.6.0
tokenizers>=0.19.0

# Regime detection
hmmlearn>=0.3.0

# Machine Learning
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# Tuning
optuna>=3.6.0

# Time series clustering
tslearn>=0.6.3

# Backtesting
lib-pybroker>=1.2.0
vectorbt>=0.26.0

# Validation
# (timeseriescv is small; install via pip install timeseriescv)
timeseriescv>=0.1

# Tracking
mlflow>=2.12.0

# LLM
openai>=1.20.0

# Broker (paper trading)
alpaca-py>=0.20.0

# Alerts
python-telegram-bot>=21.0

# Scheduling (optional, in-process)
apscheduler>=3.10.0

# Dev tools
pytest>=8.0
pytest-cov>=4.1
black>=24.0
ruff>=0.4.0
pre-commit>=3.7
ipykernel>=6.29.0
jupyter>=1.0.0
